+++ vulp/actuation/Pi3HatInterface.h
@@ -68,15 +68,13 @@ class Pi3HatInterface : public Interface {
 
   /*! Spin a new communication cycle.
    *
-   * \param[in] data Buffer to read commands from and write replies to.
    * \param[in] callback Function to call when the cycle is over.
    *
    * The callback will be invoked from an arbitrary thread when the
    * communication cycle has completed. All memory pointed to by \p data must
    * remain valid until the callback is invoked.
    */
-  void cycle(const moteus::Data& data,
-             std::function<void(const moteus::Output&)> callback) final;
+  void cycle(std::function<void(const moteus::Output&)> callback) final;
 
  private:
   /*! Main loop of the CAN thread.